# core-common-kycb-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-kycb-mgmt
  description: KYC/B & AML Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/kycb
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/compliance/parties/{partyId}/aml-screenings:
    get:
      tags:
        - AML Screening
      summary: List AML screenings
      description: Retrieves all AML screenings for the specified party ID with filtering capabilities
      operationId: listAmlScreenings
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.partyId
          in: query
          description: Exact filter for party id
          required: false
          schema:
            type: string
        - name: filters.screeningDate
          in: query
          description: Exact filter for screening date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[screeningDate].from
          in: query
          description: Filter screening date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[screeningDate].to
          in: query
          description: Filter screening date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.screeningType
          in: query
          description: Exact filter for screening type
          required: false
          schema:
            type: string
        - name: filters.matchesFound
          in: query
          description: Exact filter for matches found
          required: false
          schema:
            type: boolean
        - name: filters.matchCount
          in: query
          description: Exact filter for match count
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[matchCount].from
          in: query
          description: Filter match count from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[matchCount].to
          in: query
          description: Filter match count to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.screeningProvider
          in: query
          description: Exact filter for screening provider
          required: false
          schema:
            type: string
        - name: filters.referenceId
          in: query
          description: Exact filter for reference id
          required: false
          schema:
            type: string
        - name: filters.screeningResult
          in: query
          description: Exact filter for screening result
          required: false
          schema:
            type: string
        - name: filters.nextScreeningDate
          in: query
          description: Exact filter for next screening date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[nextScreeningDate].from
          in: query
          description: Filter next screening date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[nextScreeningDate].to
          in: query
          description: Filter next screening date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateCreated
          in: query
          description: Exact filter for date created
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].from
          in: query
          description: Filter date created from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].to
          in: query
          description: Filter date created to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateUpdated
          in: query
          description: Exact filter for date updated
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].from
          in: query
          description: Filter date updated from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].to
          in: query
          description: Filter date updated to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successfully retrieved AML screenings
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
    post:
      tags:
        - AML Screening
      summary: Create AML screening
      description: Creates a new AML screening for the specified party
      operationId: createAmlScreening
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmlScreeningDTO'
        required: true
      responses:
        '201':
          description: Successfully created AML screening
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlScreeningDTO'
...
  /api/v1/compliance/parties/{partyId}/aml-screenings/{screeningId}:
    get:
      tags:
        - AML Screening
      summary: Get screening details
      description: Retrieves a specific AML screening by its ID
      operationId: getAmlScreening
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: screeningId
          in: path
          description: ID of the screening
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved AML screening
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlScreeningDTO'
        '404':
          description: AML screening not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlScreeningDTO'
    delete:
      tags:
        - AML Screening
      summary: Delete AML screening
      description: Deletes an AML screening
      operationId: deleteAmlScreening
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: screeningId
          in: path
          description: ID of the screening
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successfully deleted AML screening
        '404':
          description: AML screening not found
    patch:
      tags:
        - AML Screening
      summary: Update AML screening
      description: Updates an existing AML screening
      operationId: updateAmlScreening
      parameters:
        - name: partyId
          in: path
          description: ID of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: screeningId
          in: path
          description: ID of the screening
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmlScreeningDTO'
        required: true
      responses:
        '200':
          description: Successfully updated AML screening
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlScreeningDTO'
        '404':
          description: AML screening not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlScreeningDTO'
...
components:
  schemas:
...
    AmlScreeningDTO:
      required:
        - matchesFound
        - partyId
        - screeningDate
        - screeningProvider
        - screeningResult
        - screeningType
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        amlScreeningId:
          type: string
          format: uuid
          readOnly: true
        partyId:
          type: string
          format: uuid
        screeningDate:
          type: string
          format: date-time
        screeningType:
          maxLength: 50
          minLength: 0
          type: string
        matchesFound:
          type: boolean
        matchCount:
          minimum: 0
          type: integer
          format: int32
        screeningProvider:
          maxLength: 100
          minLength: 0
          type: string
        referenceId:
          maxLength: 100
          minLength: 0
          type: string
        screeningResult:
          maxLength: 50
          minLength: 0
          type: string
        nextScreeningDate:
          type: string
          format: date-time
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.